Use `GB` instead of `UK` on disclaimers content by country
